move `RwError` into `frontend` crate
